backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
supabase==2.0.0
python-dotenv==1.0.0
opencv-python>=4.8.0
tensorflow>=2.15.0
numpy<2.0,>=1.22.0
Pillow>=10.2.0
gunicorn==22.0.0
